Add not owned file checker
This check is enabled when the `EXPERIMENTAL_CHECKS="notowned"` is specified.
You can also define `NOT_OWNED_CHECKER_SKIP_PATTERNS="*"`. As a result, the `*` pattern
from the CODEOWNERS file will be ignored and you will see not owned files unless
a later match takes precedence. It's useful because often we have default owners entry

```
*       @global-owner1 @global-owner2
```

but this should be a rare case that default owners are assigned.