Add not owned file checker (#8)
This check is enabled when the `EXPERIMENTAL_CHECKS="notowned"` is specified.
For example, you can specify `NOT_OWNED_CHECKER_SKIP_PATTERNS="*"` and as a result, the `*` pattern from the **CODEOWNERS** file will be ignored and files owned by this pattern will be reported as unowned unless a later specific pattern will match that path. It's useful because often we have default owners entry

```
*       @global-owner1 @global-owner2
```

and this should be a rare case that default owners are assigned.