Fix file exists checker
The main problem was that the file exists checker uses go's built in filepath globber, which doesn't support double-asterisk patterns (golang/go#11862).
Additionally, with patterns like `*.js` we need to search for all JS files not only on the first level but also in nested directories.

Based on made investigation this problem was fixed and additional test covereage was added to document that contract.